Disable the multi-pane view for now
There's a bug when you open a profile, discard, and then open a profile again. If you do this, then the view pager fragments are empty. I believe this will be fixed if the view pager uses a child fragment manager, but I can't do this until I migrate off of Sherlock.